requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
pydantic
passlib[bcrypt]
flask